Use the `Perm` class rather than `Map[Int, Int]` for permutation instances (#794)
* Use `Perm` in place of `Map[Int, Int]` for representing permutations

- Move `Perm` from laws to optional
- Define group typeclass instances against `Perm` rather than `Map`
- Move instances to `Perm` companion object
- Rename `Perm#map` to `mapping` to avoid confusion
- Rename `mapIntIntPermutation.scala` to `Perm.scala`
- Include tests for Perm group and group action laws

* Make `Perm` extend `(Int => Int)`

- adds an `apply` method that replaces calls of `perm.mappping.getOrElse(n, n)`
- adds smart `apply` constructor to enforce constraints for `Perm`
- adds a cycle-based constructor

* Add a `permute` method to `Perm` for permuting a `SeqLike`

Also adds `image` method for convenience

`PermSeqPartialAction` is now implement in terms of `Perm#permute`
Note that checking the lower bound of moved point is unnecessary
because image is constrained to non-negative indices.

* Add memoized `invert` and `inverse` methods for `Perm`

This also alows us to make `mapping` private in `Perm`, and implement
the right action in terms of `invert`

* Overload `compose` and `andThen` combinators for `Perm`

Overloads methods from `Function1` for use with another `Perm`

`compose` uses private constructor to skip constraint checks

group `combine` is now implemented in terms of `andThen`

* Add documentation to `Perm`

* Make `Perm#apply(k: Int): Int` the left action

Function application syntax, `p(n)` is the left action. Since this is
the default action of Perm on Int, `combine` should be implemented in
terms of `compose` rather than `andThen`.

Given an intetger `n`, the left group action is denoted `p·n`, so that
`(pr)⋅x == p·(r·n)` for all permutations `p` and `r`.

The difference between left and right actions is in the order in which
a product like gh acts on x. For a left action h acts first and is
followed by g, while for a right action g acts first and is followed
by h.

`Perm#permute` is also corrected so that

```
Perm(0, 2, 3).permute((1 to 5).toSeq) == Seq(4, 2, 0, 3, 5)
```

The cycle constructor now requires `cycles` that are disjoint.  Choice
of `compose` or `andThen` no longer matters, since multiplication of
disjoint cycles is commutative.